+++ src/test/kotlin/com/example/kotlintddsample/product/ProductServiceTest.kt
@@ -1,8 +1,5 @@
-package com.example.kotlintddsample
+package com.example.kotlintddsample.product
 
-import com.example.kotlintddsample.product.Product
-import com.example.kotlintddsample.product.ProductRepository
-import com.example.kotlintddsample.product.ProductService
 import com.ninjasquad.springmockk.MockkBean
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.collections.shouldHaveSize
